+++ packages/pluggableWidgets/calendar-web/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- We added a Year view, showing all 12 months at a glance with event indicators. You can choose which view opens when a day is clicked (Day, Week, Work week, Month, or Agenda).
+
 ## [2.5.0] - 2026-07-28
 
 ### Added